Update PredictiveThreshold: develop -> 1.1.1
An update to PredictiveThreshold's develop branch has made it no longer
compatible with Zenoss 5.1. Fortunately prior to this there was no
difference between PredicticeThreshold's develop branch and the 1.1.1
release tag. So by changing the ref to 1.1.1 we get back to exactly
where we were before the breaking change to PredictiveThreshold.

zenoss/ZenPacks.zenoss.PredictiveThreshold@8b11e04