Update CiscoUCS ZenPack: 2.4.3 to hotfix/2.4.4
This change is a temporary measure to allow migration testing to be
performed and validate that the fix is correct. As soon as migration is
successful, a CiscoUCS 2.4.4 release will be made, and the manifest will
be pointed to that tag instead of the hotfix/2.4.4 head.

Changes from 2.4.3 to hotfix/2.4.4:

- Fix migration issue related to ucsLayer2Catalog.

https://github.com/zenoss/ZenPacks.zenoss.CiscoUCS/compare/2.4.3...hotfix/2.4.4