Add EnterpriseLinux 1.5.0
EnterpriseLinux was removed in 472746c because it was deprecated by
LinuxMonitor 2.0.0. However, this has caused a problem when migrating
UCS-PM v1 installations to UCS-PM v2. For this reason we need to
continue shipping EnterpriseLinux in support/5.1.x. However, we can't
ship version 1.4.0 because it caused ZEN-24385 and ZEN-24388.

EnterpriseLinux 1.5.0 is a shell of a ZenPack that contains no
functionality beyond what is necessary to fix ZEN-24412. All Linux
monitoring functionality is now in the LinuxMonitor ZenPack.

This is the first step in fixing ZEN-24412.
zenoss/platform-build@support/5.1.x will also have to have its
pkg/zenoss_ucspm_zenpacks.mk updated to complete the fix.

Refs ZEN-24412.